refactor(mantaray): dissolve nectar-loadsave into mantaray::persist
Move NodeLoadSaver, LoadError, the save-error narrowing and the recording getter into crates/mantaray/src/persist.rs behind the manifest feature, beside the NodeLoader and NodeSaver seam they implement. Delete the nectar-loadsave crate, its workspace dependency entry and its lockfile records, and repoint every consumer (integration tests, benches, doctests) at the nectar-mantaray re-exports. Add MantarayManifest::over(store) for the common pass-the-same-store-twice construction, with an integration test that drives both seams through one shared store.

Closes #655.

AI Assistance: implemented by Claude Code with claude-fable-5 from the issue #655 specification; the maintainer set the scope and reviewed the result.
