+++ crates/integration-tests/Cargo.toml
@@ -32,7 +32,6 @@ bytes.workspace = true
 futures.workspace = true
 nectar-file = { workspace = true, features = ["std"] }
 nectar-ldb = { workspace = true, features = ["arbitrary", "manifest"] }
-nectar-loadsave.workspace = true
 nectar-manifest.workspace = true
 nectar-mantaray = { workspace = true, features = ["manifest"] }
 nectar-postage = { workspace = true, features = ["std"] }
@@ -49,7 +48,6 @@ thiserror = { workspace = true, features = ["std"] }
 encryption = [
 	"nectar-file/encryption",
 	"nectar-ldb/encryption",
-	"nectar-loadsave/encryption",
 	"nectar-mantaray/encryption",
 	"nectar-primitives/encryption",
 ]
