+++ crates/integration-tests/tests/manifest/common/mod.rs
@@ -6,11 +6,10 @@ use std::sync::Arc;
 
 use nectar_file::{DataSink, File, MemSink, Policy};
 use nectar_ldb::Database;
-use nectar_loadsave::NodeLoadSaver;
 use nectar_manifest::{
     Batch, Manifest, ManifestCursor, ManifestError, ManifestPath, ManifestView, MapEntry,
 };
-use nectar_mantaray::MantarayManifest;
+use nectar_mantaray::{MantarayManifest, NodeLoadSaver};
 use nectar_primitives::store::{ContentGet, MemoryStore};
 use nectar_primitives::{ChunkAddress, ChunkRef, DEFAULT_BODY_SIZE, StandardChunkSet};
