+++ crates/integration-tests/tests/manifest/policy.rs
@@ -9,9 +9,8 @@ use std::task::Poll;
 
 use nectar_file::{Policy, Window};
 use nectar_ldb::Database;
-use nectar_loadsave::NodeLoadSaver;
 use nectar_manifest::{Manifest, ManifestPath, ManifestView, MemSink};
-use nectar_mantaray::MantarayManifest;
+use nectar_mantaray::{MantarayManifest, NodeLoadSaver};
 use nectar_primitives::store::{ChunkGet, ChunkPut};
 use nectar_primitives::{
     Chunk, ChunkAddress, ChunkRef, ContentOnlyChunkSet, DEFAULT_BODY_SIZE, Verified,
