+++ crates/integration-tests/tests/manifest/root_config.rs
@@ -9,13 +9,14 @@ use std::ops::Bound;
 use std::sync::Arc;
 
 use bytes::Bytes;
-use nectar_ldb::{Builder, Database, Entry, Key, Plaintext, Reader as LdbReader, Served, V1, Website};
-use nectar_loadsave::NodeLoadSaver;
+use nectar_ldb::{
+    Builder, Database, Entry, Key, Plaintext, Reader as LdbReader, Served, V1, Website,
+};
 use nectar_manifest::{
     Batch, ListEntry, Manifest, ManifestCursor, ManifestError, ManifestMeta, ManifestOp,
     ManifestPath, ManifestView, MapEntry, MemSink, MetadataView, WellKnownKey,
 };
-use nectar_mantaray::{ManifestEditor, MantarayManifest};
+use nectar_mantaray::{ManifestEditor, MantarayManifest, NodeLoadSaver};
 use nectar_primitives::{ChunkRef, DEFAULT_BODY_SIZE, EncryptedChunkRef, EncryptionKey};
 use nectar_testing::run;
