+++ crates/integration-tests/tests/mantaray/bee_layout.rs
@@ -17,9 +17,8 @@
 
 use std::sync::Arc;
 
-use nectar_loadsave::NodeLoadSaver;
 use nectar_manifest::{Batch, Manifest, ManifestCursor, ManifestPath, ManifestView};
-use nectar_mantaray::{ManifestEditor, MantarayManifest, Reader, metadata};
+use nectar_mantaray::{ManifestEditor, MantarayManifest, NodeLoadSaver, Reader, metadata};
 use nectar_primitives::store::{ContentGet, MemoryStore};
 use nectar_primitives::{ChunkAddress, ChunkRef, DEFAULT_BODY_SIZE, StandardChunkSet};
 use nectar_testing::run;
