+++ crates/integration-tests/tests/mantaray/bmt_nodes.rs
@@ -9,8 +9,7 @@ use std::collections::BTreeSet;
 
 use bytes::Bytes;
 use nectar_file::{File, Policy};
-use nectar_loadsave::NodeLoadSaver;
-use nectar_mantaray::{AddressStream, Cursor, ManifestEditor, NodeLoader, Reader};
+use nectar_mantaray::{AddressStream, Cursor, ManifestEditor, NodeLoadSaver, NodeLoader, Reader};
 use nectar_primitives::chunk::{ChunkAddress, ChunkOps, ContentChunk};
 use nectar_primitives::store::{ContentGet, MemoryStore};
 use nectar_primitives::{EntryRef, StandardChunkSet};
