--- crates/loadsave/Cargo.toml
@@ -1,40 +1,0 @@
-[package]
-name = "nectar-loadsave"
-version.workspace = true
-edition.workspace = true
-rust-version.workspace = true
-authors.workspace = true
-license.workspace = true
-homepage.workspace = true
-repository.workspace = true
-exclude.workspace = true
-description = "Manifest node persistence over the nectar file pipeline"
-documentation = "https://docs.rs/nectar-loadsave"
-readme = "README.md"
-keywords = ["swarm", "ethereum", "manifest", "mantaray", "file"]
-categories = ["data-structures", "asynchronous"]
-
-[lints]
-workspace = true
-
-[dependencies]
-nectar-file = { workspace = true, features = ["std"] }
-nectar-mantaray.workspace = true
-nectar-primitives = { workspace = true, features = ["std"] }
-
-[dev-dependencies]
-bytes.workspace = true
-nectar-testing.workspace = true
-
-[features]
-# Encrypted node persistence through the file pipeline's encrypted split
-# mode.
-encryption = [
-	"nectar-file/encryption",
-	"nectar-mantaray/encryption",
-	"nectar-primitives/encryption",
-]
-
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "docsrs"]
