--- crates/loadsave/LICENSE
@@ -1,1 +1,0 @@
-../../LICENSE
